Turning BaseInput into a Common module
